feat: review and fix SEO metadata for all public pages
- Convert resume page to server/client split for metadata support
- Add metadata to resume page (title + description)
- All other pages (home, career, education, projects) already had proper metadata
- Dynamic pages (project detail, resume detail) already had generateMetadata
