+++ package.json
@@ -44,6 +44,7 @@
     "lucide-react": "^0.475.0",
     "mermaid": "^11.16.0",
     "mongodb": "^7.5.0",
+    "mongoose": "^9.7.4",
     "next": "^16.2.10",
     "next-themes": "^0.4.4",
     "qrcode.react": "^4.2.0",
